feat(auth): Refactored user data payload creation
