+++ server/src/global/auth/authController.ts
@@ -6,8 +6,8 @@ import { CookieOptions, Request, Response } from "express";
 import database = require("../../database");
 import * as globalTypes from "./../globalTypes";
 import * as authTypes from "./authTypes";
-import { doesPermissionsHaveSomeOf, doesPermissionsInclude } from "../../utils/authUtils";
 import { generateJWT, verifyJWT } from "../../utils/jwtUtils";
+import { createUserDataPayload } from "../../utils/authUtils";
 
 /**
  * Options for the cookie
@@ -48,7 +48,7 @@ export const login = (req: Request, res: Response): void => {
         GROUP BY mitgliedID, name`,
         [req.body.username]
       )
-      .then((result: authTypes.LoginQueryResult[]) => {
+      .then((result: authTypes.UserDataQueryResult[]) => {
         if (result.length === 0) {
           // Sleeping randomly between 50 and 100 miliseconds to prevent username prediction
           sleepRandomly(50, 110);
@@ -67,38 +67,11 @@ export const login = (req: Request, res: Response): void => {
             [result[0].mitgliedID]
           )
           .then((directorPermissionsResult: globalTypes.Permission[]) => {
-            let permissions = [];
-            let roles = [];
-            // Adds role permissions to the permissions array and adds directorID to the roles array
-            if (directorPermissionsResult.length !== 0) {
-              directorPermissionsResult.forEach((permission) => {
-                permissions.push({ permissionID: permission.permissionID, canDelegate: permission.canDelegate });
-                if (!roles.includes(permission.directorID)) {
-                  roles.push(permission.directorID);
-                }
-              });
-            }
-
-            // Adds normal permissions to the permissions array
-            if (result[0].permissions) {
-              result[0].permissions
-                .split(",")
-                .map(Number)
-                .map((perm) => {
-                  // A Permission which was delegated to a member cannot be delegated further (therefore canDelegate is always 0)
-                  permissions.push({ permissionID: perm, canDelegate: 0 });
-                });
-            }
             bcrypt
               .compare(req.body.password, result[0].passwordHash)
               .then((match) => {
                 if (match) {
-                  const payload: globalTypes.JWTPayload = {
-                    mitgliedID: result[0].mitgliedID,
-                    name: result[0].name,
-                    permissions,
-                    roles,
-                  };
+                  const payload = createUserDataPayload(result[0], directorPermissionsResult);
                   const token = generateJWT(payload);
                   res.cookie("token", token, cookieOptions).status(200).json(payload);
                 } else {
@@ -139,7 +112,7 @@ export const retrieveUserData = (req: Request, res: Response) => {
       // Selects permissions belonging to a possible role of the member
       database
         .query(
-          `SELECT berechtigung_berechtigungID AS permissionID, canDelegate
+          `SELECT berechtigung_berechtigungID AS permissionID, canDelegate, mitglied_has_evposten.evposten_evpostenID as directorID
           FROM mitglied_has_evposten
           LEFT JOIN evposten_has_berechtigung ON mitglied_has_evposten.evposten_evpostenID = evposten_has_berechtigung.evposten_evpostenID
           WHERE mitglied_has_evposten.mitglied_mitgliedID = ?;
@@ -147,24 +120,8 @@ export const retrieveUserData = (req: Request, res: Response) => {
           [jwtData.mitgliedID]
         )
         .then((directorPermissionsResult: globalTypes.Permission[]) => {
-          let permissions = [];
-
-          // Adds role permissions to the permissions array
-          if (directorPermissionsResult.length !== 0) {
-            permissions = directorPermissionsResult;
-          }
-
-          // Adds normal permissions to the permissions array
-          if (result[0].permissions) {
-            result[0].permissions
-              .split(",")
-              .map(Number)
-              .map((perm) => {
-                // A Permission which was delegated to a member cannot be delegated further (therefore canDelegate is always 0)
-                permissions.push({ permissionID: perm, canDelegate: 0 });
-              });
-          }
-          res.status(200).json({ ...result[0], permissions });
+          const payload = createUserDataPayload(result[0], directorPermissionsResult);
+          res.status(200).json(payload);
         })
         .catch(() => {
           res.status(500).send("Query Error");
