+++ server/src/global/auth/authTypes.ts
@@ -3,21 +3,12 @@
  */
 
 /**
- * Type of query results of the login
- */
-export type LoginQueryResult = {
-  mitgliedID: number;
-  name: string;
-  passwordHash: string;
-  permissions: string;
-};
-
-/**
  * Type of query results of the user data retrieval
  */
 export type UserDataQueryResult = {
-  mitliedID: number;
+  mitgliedID: number;
   name: string;
+  passwordHash?: string;
   permissions: string;
 };
